requirements.txt
pypdf>=3.0.0
reportlab>=3.6.0
cryptography>=3.1
PyMuPDF>=1.23.0
Pillow>=9.0.0
